# docs/openapi/cliser-data-node.openapi.current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLISER DATA NODE API",
    "version": "1.0.0",
    "description": "HTTP API do CLISER DATA NODE.\n\nAPI Version: v1\nBase Path: /api/v1"
  },
  "paths": {
...
    "/api/v1/namespaces/{namespace}/objects": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "List Objects",
        "operationId": "list_objects_api_v1_namespaces__namespace__objects_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListResponse"
                },
                "examples": {
                  "ObjectListResponse": {
                    "$ref": "#/components/examples/ObjectListResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "objects"
        ],
        "summary": "Create Object",
        "operationId": "create_object_api_v1_namespaces__namespace__objects_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            },
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectCreateRequest"
              },
              "examples": {
                "ObjectCreate": {
                  "$ref": "#/components/examples/ObjectCreateRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectCreateResponse"
                },
                "examples": {
                  "ObjectCreateResponse": {
                    "$ref": "#/components/examples/ObjectCreateResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}/objects/{object_id}": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Get Object",
        "operationId": "get_object_api_v1_namespaces__namespace__objects__object_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectMetadataResponse"
                },
                "examples": {
                  "ObjectMetadataResponse": {
                    "$ref": "#/components/examples/ObjectMetadataResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "objects"
        ],
        "summary": "Delete Object",
        "operationId": "delete_object_api_v1_namespaces__namespace__objects__object_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectDeleteResponse"
                },
                "examples": {
                  "ObjectDeleteResponse": {
                    "$ref": "#/components/examples/ObjectDeleteResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}/objects/{object_id}/verify": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "Verify Object",
        "operationId": "verify_object_api_v1_namespaces__namespace__objects__object_id__verify_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectVerifyResponse"
                },
                "examples": {
                  "ObjectVerifyResponse": {
                    "$ref": "#/components/examples/ObjectVerifyResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace}/quota": {
      "get": {
        "tags": [
          "quotas"
        ],
        "summary": "Get Quota",
        "operationId": "get_quota_api_v1_namespaces__namespace__quota_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                },
                "examples": {
                  "QuotaResponse": {
                    "$ref": "#/components/examples/QuotaResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "quotas"
        ],
        "summary": "Update Quota",
        "operationId": "update_quota_api_v1_namespaces__namespace__quota_put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuotaUpdateRequest"
              },
              "examples": {
                "QuotaUpdate": {
                  "$ref": "#/components/examples/QuotaUpdateRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaResponse"
                },
                "examples": {
                  "QuotaResponse": {
                    "$ref": "#/components/examples/QuotaResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ObjectCreateRequest": {
        "properties": {
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ObjectCreateRequest"
      },
      "ObjectCreateResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "content_hash",
          "status"
        ],
        "title": "ObjectCreateResponse"
      },
      "ObjectDeleteResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "deleted",
          "status"
        ],
        "title": "ObjectDeleteResponse"
      },
      "ObjectListItem": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "status"
        ],
        "title": "ObjectListItem"
      },
      "ObjectListResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectListItem"
            },
            "type": "array",
            "title": "Objects"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "objects",
          "count"
        ],
        "title": "ObjectListResponse"
      },
      "ObjectMetadataResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "content_hash",
          "status"
        ],
        "title": "ObjectMetadataResponse"
      },
      "ObjectVerifyResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "valid",
          "status",
          "message"
        ],
        "title": "ObjectVerifyResponse"
      },
      "QuotaResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "quota_bytes": {
            "type": "integer",
            "title": "Quota Bytes"
          },
          "used_bytes": {
            "type": "integer",
            "title": "Used Bytes"
          },
          "available_bytes": {
            "type": "integer",
            "title": "Available Bytes"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "quota_bytes",
          "used_bytes"
        ],
        "title": "QuotaResponse"
      },
      "QuotaUpdateRequest": {
        "properties": {
          "quota_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quota Bytes"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "LOGICAL"
          }
        },
        "type": "object",
        "required": [
          "quota_bytes"
        ],
        "title": "QuotaUpdateRequest"
      },
...
      "APIError": {
        "title": "APIError",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "OBJECT_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Objeto não encontrado"
              },
              "request_id": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "req_01JEXAMPLE"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "credential",
        "description": "Credencial de API enviada no cabeçalho Authorization como Bearer <credential>."
      }
    },
    "examples": {
...
      "ObjectCreateRequest": {
        "value": {
          "data": "Q0xJU0VSIERBVEEgTk9ERQ=="
        },
        "summary": "Objeto em Base64",
        "description": "O campo data deve conter Base64 válido."
      },
      "ObjectCreateResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "size": 16,
          "content_hash": "sha256:01JEXAMPLE",
          "status": "ACTIVE"
        },
        "summary": "Objeto criado"
      },
      "ObjectListResponse": {
        "value": {
          "namespace": "documents",
          "objects": [
            {
              "object_id": "obj_01JEXAMPLE",
              "namespace": "documents",
              "size": 16,
              "created_at": "2026-09-14T12:00:00Z",
              "status": "ACTIVE"
            }
          ],
          "count": 1
        },
        "summary": "Lista de objetos"
      },
      "ObjectMetadataResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "size": 16,
          "content_hash": "sha256:01JEXAMPLE",
          "created_at": "2026-09-15T15:00:00Z",
          "status": "ACTIVE"
        },
        "summary": "Metadados do objeto"
      },
      "ObjectVerifyResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "valid": true,
          "status": "VALID",
          "message": "VALID"
        },
        "summary": "Objeto válido"
      },
      "ObjectDeleteResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "deleted": true,
          "status": "DELETED"
        },
        "summary": "Objeto removido"
      },
...
      "QuotaUpdateRequest": {
        "value": {
          "quota_bytes": 10737418240,
          "mode": "LOGICAL"
        },
        "summary": "Atualização de quota"
      },
      "QuotaResponse": {
        "value": {
          "namespace": "documents",
          "quota_bytes": 10737418240,
          "used_bytes": 245760,
          "available_bytes": 10737172480
        },
        "summary": "Quota do namespace"
      },
...
      "ErrorValidation": {
        "value": {
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Request validation failed.",
            "request_id": "req_01JEXAMPLE",
            "details": {
              "errors": [
                {
                  "type": "value_error",
                  "loc": [
                    "body",
                    "namespace"
                  ],
                  "msg": "field required"
                }
              ]
            }
          }
        },
        "summary": "Erro de validação"
      },
...
    }
  },
...
}